# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_JobDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobDetail"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data"
        ],
        "title": "ApiResponse[JobDetail]",
        "type": "object"
      },
...
      "ApiResponse_dict_": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data"
        ],
        "title": "ApiResponse[dict]",
        "type": "object"
      },
...
      "ApiResponse_list_QueueDepth__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/QueueDepth"
            },
            "title": "Data",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data"
        ],
        "title": "ApiResponse[list[QueueDepth]]",
        "type": "object"
      },
...
      "ApiResponse_list_WorkerInfo__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/WorkerInfo"
            },
            "title": "Data",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data"
        ],
        "title": "ApiResponse[list[WorkerInfo]]",
        "type": "object"
      },
      "ApiResponse_list_dict__": {
        "properties": {
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Data",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data"
        ],
        "title": "ApiResponse[list[dict]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobDetail": {
        "properties": {
          "args": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Args"
          },
          "checks_completed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checks Completed"
          },
          "checks_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checks Total"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "findings_fail": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Findings Fail"
          },
          "findings_found": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Findings Found"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kwargs": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kwargs"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "title": "Logs",
            "type": "array"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "retries": {
            "default": 0,
            "title": "Retries",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_name": {
            "title": "Task Name",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "traceback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traceback"
          },
          "worker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker"
          }
        },
        "required": [
          "job_id",
          "task_name",
          "tenant_id",
          "status"
        ],
        "title": "JobDetail",
        "type": "object"
      },
...
      "Meta": {
        "properties": {
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "title": "Meta",
        "type": "object"
      },
...
      "QueueDepth": {
        "properties": {
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "queue": {
            "title": "Queue",
            "type": "string"
          }
        },
        "required": [
          "queue",
          "depth"
        ],
        "title": "QueueDepth",
        "type": "object"
      },
...
      "RetryRequest": {
        "properties": {
          "actor_email": {
            "title": "Actor Email",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "actor_email",
          "tenant_id"
        ],
        "title": "RetryRequest",
        "type": "object"
      },
...
      "TaskType": {
        "enum": [
          "cspm",
          "iac",
          "discovery"
        ],
        "title": "TaskType",
        "type": "string"
      },
      "TriggerRequest": {
        "properties": {
          "actor_email": {
            "title": "Actor Email",
            "type": "string"
          },
          "frameworks": {
            "items": {
              "type": "string"
            },
            "title": "Frameworks",
            "type": "array"
          },
          "provider": {
            "default": "",
            "title": "Provider",
            "type": "string"
          },
          "provider_id": {
            "default": "",
            "title": "Provider Id",
            "type": "string"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "task_type",
          "actor_email"
        ],
        "title": "TriggerRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkerInfo": {
        "properties": {
          "active_tasks": {
            "default": 0,
            "title": "Active Tasks",
            "type": "integer"
          },
          "hostname": {
            "title": "Hostname",
            "type": "string"
          },
          "processed": {
            "default": 0,
            "title": "Processed",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "uptime_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uptime Seconds"
          }
        },
        "required": [
          "hostname",
          "status"
        ],
        "title": "WorkerInfo",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Open CNAPP \u2014 Built for Everyone",
    "title": "Ogum Security API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/jobs": {
      "get": {
        "description": "List scan jobs across all tenants or scoped to a specific tenant.",
        "operationId": "list_jobs_api_v1_admin_jobs_get",
        "parameters": [
          {
            "description": "Filter to a specific tenant. Omit for all tenants.",
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to a specific tenant. Omit for all tenants.",
              "title": "Tenant Id"
            }
          },
          {
            "description": "Filter by job status (e.g. queued, running, failed)",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by job status (e.g. queued, running, failed)",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Keyset cursor (created_at ISO timestamp of last item)",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Keyset cursor (created_at ISO timestamp of last item)",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Jobs",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/jobs/trigger": {
      "post": {
        "description": "Immediately dispatch a new scan job, bypassing the scheduled queue.",
        "operationId": "trigger_job_api_v1_admin_jobs_trigger_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trigger Job",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/jobs/{job_id}": {
      "delete": {
        "description": "Cancel a queued or running job.",
        "operationId": "revoke_job_api_v1_admin_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor_email",
            "required": true,
            "schema": {
              "title": "Actor Email",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Job",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "description": "Get full detail for a single scan job.",
        "operationId": "get_job_api_v1_admin_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "description": "Tenant that owns this job",
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "description": "Tenant that owns this job",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_JobDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Job",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/jobs/{job_id}/logs": {
      "get": {
        "description": "Stream stored log lines for a job via Server-Sent Events.",
        "operationId": "stream_job_logs_api_v1_admin_jobs__job_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "description": "Tenant that owns this job",
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "description": "Tenant that owns this job",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stream Job Logs",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/jobs/{job_id}/retry": {
      "post": {
        "description": "Re-enqueue a failed job. Returns the new job_id.",
        "operationId": "retry_job_api_v1_admin_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Job",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/queue-depth": {
      "get": {
        "description": "Return the number of pending messages in each Celery queue.",
        "operationId": "get_queue_depth_api_v1_admin_queue_depth_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_QueueDepth__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Queue Depth",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/tenants": {
      "get": {
        "description": "Lista os tenants registrados (sem expor hashes de token).",
        "operationId": "list_registered_tenants_api_v1_admin_tenants_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_dict__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Registered Tenants",
        "tags": [
          "admin-tenants"
        ]
      }
    },
    "/api/v1/admin/tenants/{tenant_id}/api-token": {
      "put": {
        "description": "Registra o tenant (se necess\u00e1rio) e emite/rotaciona o token API.\n\nO valor do token \u00e9 retornado **uma \u00fanica vez** nesta resposta; o registro\npersiste apenas o hash. Emitir novamente revoga o token anterior.",
        "operationId": "issue_api_token_api_v1_admin_tenants__tenant_id__api_token_put",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform_admin",
            "required": false,
            "schema": {
              "default": false,
              "title": "Platform Admin",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Issue Api Token",
        "tags": [
          "admin-tenants"
        ]
      }
    },
    "/api/v1/admin/workers": {
      "get": {
        "description": "Return info about active Celery workers.",
        "operationId": "list_workers_api_v1_admin_workers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_WorkerInfo__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Workers",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}